Normalise permission bits to 755 or 644
Apparently, the only permission bit git tracks is executable or not. So
this flattens the permission bits to 755 for executable files, or 644
for non-executable, matching the default umask on Debian systems.

I still want to think about this:

- Does wheel installation preserve any more permission bits? If not,
this is definitely fine.
- Do other version control systems store more permissions?
- Is there a reason you might want other permissions on files installed
from wheels, e.g. to make a file not world readable? It seems redundant
when packages come from a public index, but maybe for privately
installed packages it makes sense.