Refactor WheelBuilder not to work out its own target file
With this refactoring, WheelBuilder is no longer responsible for working
out its own target file. A new wheel_main function gives it a temporary
file to write to, and renames that afterwards (because we don't know the
final name until we've processed the metadata).

Besides general loose-coupling principles, this should make it easier
for 'flit install' to work by building a wheel and calling pip with it:
that machinery will use WheelBuilder directly to make the wheel in a
temporary location.

I'm not sure whether WheelBuilder should get:
1. The path to a file it should open and write
2. A writable file object that it should write zip data to
3. A ZipFile object that it should write into.

Currently, this does 2, because it works neatly with mkstemp. Any clear
advantages to the other options?