Only install reqs in sys.executable if needed
In most cases, we can get the docstring and version number from the AST,
so we don't need the package to be importable.
If getting that info fails with ImportError,
then we install the requirements before trying again.

Also, when we install directly (for --symlink or --pth-file),
ensure that the requirements are installed to the target environment.

Closes gh-165
Closes gh-164