Reuse wheel metadata in sdist
I was not able to tell if there is a downside to using the same
metadata as the wheel but a cursory investigation revealed
that it's what Poetry does.  setuptools omits the requirements
but its PKG-INFO appears to be otherwise identical to the wheel's
METADATA file.  flit appears to be a bit of an outlier in creating
a barebones PKG-INFO file.

Closes #409.