Fix license file duplication in sdists
This bug can happen on Windows when the `license-files` setting contains
a glob that matches one of the files in the source directory. In this case,
the same path can be returned by `select_files` both in the POSIX format and
the Windows format, which makes the deduplication logic in
`apply_includes_excludes` fail.

Fix it by making sure that all paths passed to `apply_includes_excludes` are
in the native OS format.